Add a bunch of days started/unfinished at once

+++ 2024/zig-2024/src/types.zig
@@ -1,5 +1,5 @@
 pub const Answer = union(enum) {
-    number: u32,
+    number: u64,
     string: []const u8,
 };
